pyproject.toml
[tool.poetry]
name = "altruisticfearfulsearchengine"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
replit = "^2.0.0"
requests = "^2.25.1"
python-decouple = "^3.4"
asyncio = "^3.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
